+++ fastfem/elements/linear_simplex_mesh2d.py
@@ -1,0 +1,5 @@
+from fastfem.elements.element2d import StaticElement2D
+
+
+class LinearSimplexMesh2D(StaticElement2D):
+    pass
