+++ tests/elements/test_element2d.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from fastfem.elements.element import Element2D
+from fastfem.elements.element2d import Element2D
 from fastfem.elements.linear_simplex2d import LinearSimplex2D
 from fastfem.elements.spectral_element import SpectralElement2D
 from fastfem.fields.field import Field
